게시판 마무리중

+++ src/main/java/com/smartFarm/was/web/controller/BoardController.java
@@ -1,5 +1,7 @@
 package com.smartFarm.was.web.controller;
 
+import com.smartFarm.was.domain.dto.mapping.BoardMemberMappingDto;
+import com.smartFarm.was.domain.entity.Board;
 import com.smartFarm.was.domain.entity.sub.Status;
 import com.smartFarm.was.domain.response.ResultCode;
 import com.smartFarm.was.domain.response.board.AddBoardResponse;
@@ -8,9 +10,8 @@ import com.smartFarm.was.domain.request.board.AddBoardForm;
 import com.smartFarm.was.domain.dto.board.UpdateBoardDto;
 import com.smartFarm.was.domain.request.board.UpdateBoardForm;
 import com.smartFarm.was.domain.response.ResultResponse;
-import com.smartFarm.was.domain.dto.board.BoardDetailDto;
 import com.smartFarm.was.domain.response.board.BoardResponse;
-import com.smartFarm.was.domain.entity.Member;
+import com.smartFarm.was.domain.response.board.UpdateBoardResponse;
 import com.smartFarm.was.web.service.BoardService;
 import com.smartFarm.was.web.service.CommentService;
 import com.smartFarm.was.web.utils.MemberAuthenticationUtils;
@@ -19,12 +20,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.HttpServletRequest;
 import java.sql.Timestamp;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 
 
 @Slf4j
@@ -57,15 +54,15 @@ public class BoardController {
             return new ResultResponse<>(HttpStatus.FORBIDDEN, ResultCode.FORBIDDEN.getCode(), "로그인 후 이용 가능한 서비스입니다.");
         }
 
-        Member member = MemberAuthenticationUtils.getMemberAuthentication();
+        Board board = Board.of(addBoardForm, MemberAuthenticationUtils.getMemberIdByMemberAuthentication());
 
-        long boardId = boardService.addBoard(addBoardForm, member.getMemberId());
+        long boardId = boardService.addBoard(board);
 
         return new ResultResponse(HttpStatus.OK, ResultCode.SUCCESS.getCode(), ResultCode.SUCCESS.getMessage(), new AddBoardResponse(boardId));
     }
 
-    @GetMapping("/detail/{id}")
-    public ResultResponse getBoardDetail(@PathVariable("id") long boardId) throws Exception {
+    @GetMapping("/detail/{boardId}")
+    public ResultResponse getBoardDetail(@PathVariable long boardId) throws Exception {
 
         BoardDetailResponse boardDetailResponse = boardService.getBoardDetail(boardId);
 
@@ -80,45 +77,50 @@ public class BoardController {
         return new ResultResponse<>(HttpStatus.OK, ResultCode.SUCCESS.getCode(), ResultCode.SUCCESS.getMessage(), boardDetailResponse);
     }
 
-    @GetMapping("/delete/{id}")
-    public ResultResponse removeBoard(HttpServletRequest httpServletRequest, @PathVariable("id") long boardId) throws Exception {
+    @GetMapping("/delete/{boardId}")
+    public ResultResponse deleteBoard(@PathVariable long boardId) throws Exception {
 
         if (MemberAuthenticationUtils.isAnonymous()) {
             return new ResultResponse<>(HttpStatus.FORBIDDEN, ResultCode.FORBIDDEN.getCode(), "로그인 후 이용 가능한 서비스입니다.");
         }
 
-        boardService.deleteBoard(boardId);
+        BoardMemberMappingDto boardMemberMappingDto = BoardMemberMappingDto.of(boardId, MemberAuthenticationUtils.getMemberIdByMemberAuthentication());
+
+        if(boardService.checkOwnerById(boardMemberMappingDto)) {
+            boardService.deleteBoard(boardId);
+        } else {
+            return new ResultResponse(HttpStatus.BAD_REQUEST, ResultCode.FAIL.getCode(), "잘못된 요청입니다.");
+        }
 
         return new ResultResponse(HttpStatus.OK, ResultCode.SUCCESS.getCode(), ResultCode.SUCCESS.getMessage());
     }
 
-    @PostMapping("/update/{id}")
-    public ResultResponse changeBoard(HttpServletRequest httpServletRequest,
-                                      @RequestBody UpdateBoardForm updateBoardForm,
-                                      @PathVariable("id") long boardId) throws Exception
-    {
+    @PostMapping("/update/{boardId}")
+    public ResultResponse updateBoard(@RequestBody UpdateBoardForm updateBoardForm, @PathVariable long boardId) throws Exception {
 
         if (MemberAuthenticationUtils.isAnonymous()) {
             return new ResultResponse<>(HttpStatus.FORBIDDEN, ResultCode.FORBIDDEN.getCode(), "로그인 후 이용 가능한 서비스입니다.");
         }
 
-        Member member = MemberAuthenticationUtils.getMemberAuthentication();
+        BoardMemberMappingDto boardMemberMappingDto = BoardMemberMappingDto.of(boardId, MemberAuthenticationUtils.getMemberIdByMemberAuthentication());
 
-        UpdateBoardDto updateBoardDto = UpdateBoardDto.builder()
-                .boardId(boardId)
-                .memberId(member.getMemberId())
-                .categoryId(updateBoardForm.getCategoryId())
-                .boardTitle(updateBoardForm.getBoardTitle())
-                .boardContent(updateBoardForm.getBoardContent())
-                .boardStatus(updateBoardForm.getBoardStatus())
-                .boardLastUpdated(new Timestamp(System.currentTimeMillis()))
-                .build();
+        if (boardService.checkOwnerById(boardMemberMappingDto)) {
 
-        Optional<BoardDetailDto> updatedDetail = boardService.updateBoard(updateBoardDto);
+            UpdateBoardDto updateBoardDto = UpdateBoardDto.builder()
+                    .boardId(boardId)
+                    .memberId(MemberAuthenticationUtils.getMemberIdByMemberAuthentication())
+                    .categoryId(updateBoardForm.getCategoryId())
+                    .boardTitle(updateBoardForm.getBoardTitle())
+                    .boardContent(updateBoardForm.getBoardContent())
+                    .boardStatus(updateBoardForm.getBoardStatus())
+                    .boardLastUpdated(new Timestamp(System.currentTimeMillis()))
+                    .build();
 
-        Map<String, BoardDetailDto> mapper = new HashMap<>();
-        mapper.put("detail", updatedDetail.get());
+            boardService.updateBoard(updateBoardDto);
+        } else {
+            return new ResultResponse(HttpStatus.BAD_REQUEST, ResultCode.FAIL.getCode(), "잘못된 요청입니다.");
+        }
 
-        return new ResultResponse(HttpStatus.OK, ResultCode.SUCCESS.getCode(), ResultCode.SUCCESS.getMessage(), null);
+        return new ResultResponse(HttpStatus.OK, ResultCode.SUCCESS.getCode(), ResultCode.SUCCESS.getMessage(), new UpdateBoardResponse(boardId));
     }
 }
